rename: MissionWeeklyCard로 변경하여 공통 컴포넌트화
